Update layout and interaction in game UI
The layout of two reply buttons in hello-view.fxml has been tweaked for better visibility and readablity. Button actions in GameController.java have been restructured to better handle game progress, with the introduction of an "introFinished" flag. Furthermore, a new character 'Mike' was added in the createCharacters() function with its dialogue.